+++ examples/scripts/spm_MAP.py
@@ -2,8 +2,16 @@ import numpy as np
 
 import pybop
 
-# Define model
+# Construct and update initial parameter values
 parameter_set = pybop.ParameterSet.pybamm("Chen2020")
+parameter_set.update(
+    {
+        "Negative electrode active material volume fraction": 0.63,
+        "Positive electrode active material volume fraction": 0.51,
+    }
+)
+
+# Define model
 model = pybop.lithium_ion.SPM(parameter_set=parameter_set)
 
 # Fitting parameters
@@ -12,21 +20,16 @@ parameters = pybop.Parameters(
         "Negative electrode active material volume fraction",
         prior=pybop.Gaussian(0.6, 0.05),
         bounds=[0.5, 0.8],
+        true_value=parameter_set["Negative electrode active material volume fraction"],
     ),
     pybop.Parameter(
         "Positive electrode active material volume fraction",
         prior=pybop.Gaussian(0.48, 0.05),
         bounds=[0.4, 0.7],
+        true_value=parameter_set["Positive electrode active material volume fraction"],
     ),
 )
 
-# Set initial parameter values
-parameter_set.update(
-    {
-        "Negative electrode active material volume fraction": 0.63,
-        "Positive electrode active material volume fraction": 0.51,
-    }
-)
 # Generate data
 sigma = 0.005
 t_eval = np.arange(0, 900, 3)
@@ -44,8 +47,8 @@ dataset = pybop.Dataset(
 
 # Generate problem, cost function, and optimisation class
 problem = pybop.FittingProblem(model, parameters, dataset)
-cost = pybop.MAP(problem, pybop.GaussianLogLikelihoodKnownSigma)
-optim = pybop.CMAES(
+cost = pybop.MAP(problem, pybop.GaussianLogLikelihoodKnownSigma, sigma0=sigma)
+optim = pybop.AdamW(
     cost,
     max_unchanged_iterations=20,
     min_iterations=20,
@@ -54,6 +57,7 @@ optim = pybop.CMAES(
 
 # Run the optimisation
 x, final_cost = optim.run()
+print("True parameters:", parameters.true_value())
 print("Estimated parameters:", x)
 
 # Plot the timeseries output