+++ pybop/optimisers/base_optimiser.py
@@ -160,8 +160,7 @@ class BaseOptimiser:
 
         # Store the optimised parameters
         x = self.result.x
-        if hasattr(self.cost, "parameters"):
-            self.store_optimised_parameters(x)
+        self.parameters.update(values=x)
 
         # Check if parameters are viable
         if self.physical_viability:
@@ -181,20 +180,6 @@ class BaseOptimiser:
             If the method has not been implemented by the subclass.
         """
         raise NotImplementedError
-
-    def store_optimised_parameters(self, x):
-        """
-        Update the problem parameters with optimised values.
-
-        The optimised parameter values are stored within the associated PyBOP parameter class.
-
-        Parameters
-        ----------
-        x : array-like
-            Optimised parameter values.
-        """
-        for i, param in enumerate(self.cost.parameters):
-            param.update(value=x[i])
 
     def check_optimal_parameters(self, x):
         """