+++ pybop/plotting/plot_parameters.py
@@ -1,6 +1,6 @@
 import sys
 
-from pybop import StandardSubplot
+from pybop import GaussianLogLikelihood, StandardSubplot
 
 
 def plot_parameters(optim, show=True, **layout_kwargs):
@@ -50,6 +50,10 @@ def plot_parameters(optim, show=True, **layout_kwargs):
     trace_names = parameters.keys()
     for name in trace_names:
         axis_titles.append(("Function Call", name))
+
+    if isinstance(optim.cost, GaussianLogLikelihood):
+        axis_titles.append(("Function Call", "Sigma"))
+        trace_names.append("Sigma")
 
     # Set subplot layout options
     layout_options = dict(